refactor: NacimientoViewmodel.kt, MainViewmodel.kt
