+++ terrabit_app/app/src/main/java/com/example/terrabit_app/MainActivity.kt
@@ -13,6 +13,7 @@ import com.example.terrabit_app.navegacion.Navigation
 import com.example.terrabit_app.ui.theme.Terrabit_appTheme
 import com.example.terrabit_app.viewmodel.DrawerViewModel  // ← NUEVO IMPORT
 import com.example.terrabit_app.viewmodel.MainViewmodel
+import com.example.terrabit_app.viewmodel.NacimientoViewmodel
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -26,7 +27,8 @@ class MainActivity : ComponentActivity() {
                 ) {
                     val myViewModel: MainViewmodel = viewModel()
                     val drawerViewModel: DrawerViewModel = viewModel()
-                    Navigation(myViewModel, drawerViewModel)
+                    val ViewModelnacimiento: NacimientoViewmodel = viewModel()
+                    Navigation(myViewModel, drawerViewModel, ViewModelnacimiento)
                 }
             }
         }
